feat: convert to tsx; reformatting
